패키지 구조 설정 완료
