+++ backend/src/main/java/com/musinsa/backend/controller/BrandController.java
@@ -1,8 +1,7 @@
-package com.musinsa.backend.external.controller;
+package com.musinsa.backend.controller;
 
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class SearchController {
-
+public class BrandController {
 }
