--- backend/src/main/java/com/musinsa/backend/internal/controller/BrandController.java
@@ -1,7 +1,0 @@
-package com.musinsa.backend.internal.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class BrandController {
-}
