--- backend/src/main/java/com/musinsa/backend/internal/model/Brand.java
@@ -1,4 +1,0 @@
-package com.musinsa.backend.internal.model;
-
-public class Brand {
-}
