--- backend/src/main/java/com/musinsa/backend/internal/service/BrandServiceImpl.java
@@ -1,4 +1,0 @@
-package com.musinsa.backend.internal.service;
-
-public class BrandServiceImpl {
-}
