+++ backend/src/main/java/com/musinsa/backend/mapper/BrandMapper.java
@@ -1,0 +1,4 @@
+package com.musinsa.backend.mapper;
+
+public class BrandMapper {
+}
