+++ backend/src/main/java/com/musinsa/backend/model/Brand.java
@@ -1,0 +1,4 @@
+package com.musinsa.backend.model;
+
+public class Brand {
+}
