+++ backend/src/main/java/com/musinsa/backend/service/BrandService.java
@@ -1,0 +1,4 @@
+package com.musinsa.backend.service;
+
+public interface BrandService {
+}
