+++ backend/src/main/java/com/musinsa/backend/service/BrandServiceImpl.java
@@ -1,0 +1,4 @@
+package com.musinsa.backend.service;
+
+public class BrandServiceImpl {
+}
